pyproject.toml
[project]
name = "ptcgpb-companion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pyqt6>=6.10.2",
]
